requirements.txt
git+https://github.com/Rapptz/discord.py.git
pyNaCl==1.3.0
pandas
dnspython==1.16.0
async-timeout==3.0.1
datetime
wikipedia
requests
google
newspaper3k